Rename file. Restructured code.
